INDY-1637: add optional RocksDB usage for ValidatorInfo class.
